Revert "Merge pull request #92007 from gijoe0295/feat/ai-feature-promo-modal"
This reverts commit 135a1669243e1eaffb5022f3c34d2c11bf3318a3, reversing
changes made to 5d48c1942ced4d418f4237ccfc0d672fef87f4b7.
