+++ jest/setupAfterEnv.ts
@@ -6,16 +6,6 @@ import ONYXKEYS from '@src/ONYXKEYS';
 
 jest.useRealTimers();
 
-// Globally short-circuit AIFeaturesPromoGuard in tests. The real guard proactively
-// navigates any authenticated session to /ai-features-promo unless the dismissal NVP
-// is set, which would otherwise intercept navigation in unrelated UI tests. Tests
-// that need the real guard can override this mock locally.
-jest.mock('@libs/Navigation/guards/AIFeaturesPromoGuard', () => ({
-    __esModule: true,
-    default: {name: 'AIFeaturesPromoGuard', evaluate: () => ({type: 'ALLOW'})},
-    onSessionOrLoadingAppChanged: jest.fn(),
-}));
-
 // Patch Keyboard.addListener to return a subscription object with .remove() so that
 // @react-navigation/bottom-tabs useIsKeyboardShown hook doesn't crash on cleanup.
 if (Keyboard && typeof Keyboard.addListener === 'function') {
