+++ src/SCREENS.ts
@@ -970,10 +970,6 @@ const SCREENS = {
         ROOT: 'MigratedUserWelcomeModal_Root',
     },
 
-    AI_FEATURES_PROMO_MODAL: {
-        ROOT: 'AIFeaturesPromoModal_Root',
-    },
-
     TEST_DRIVE_MODAL: {
         ROOT: 'TestDrive_Modal_Root',
     },
