--- src/components/FeatureTrainingModal/index.tsx
@@ -1,427 +1,0 @@
-import type {ImageContentFit} from 'expo-image';
-import React, {useEffect, useRef, useState} from 'react';
-import {View} from 'react-native';
-// eslint-disable-next-line no-restricted-imports
-import type {LayoutChangeEvent, ScrollView as RNScrollView, StyleProp, TextStyle, ViewStyle} from 'react-native';
-import type {MergeExclusive} from 'type-fest';
-import type ImageSVGProps from '@components/ImageSVG/types';
-import type DotLottieAnimation from '@components/LottieAnimations/types';
-import Modal from '@components/Modal';
-import ScrollView from '@components/ScrollView';
-import useKeyboardState from '@hooks/useKeyboardState';
-import useResponsiveLayout from '@hooks/useResponsiveLayout';
-import useSafeAreaInsets from '@hooks/useSafeAreaInsets';
-import useStyleUtils from '@hooks/useStyleUtils';
-import useThemeStyles from '@hooks/useThemeStyles';
-import useWindowDimensions from '@hooks/useWindowDimensions';
-import isInLandscapeModeUtil from '@libs/isInLandscapeMode';
-import Log from '@libs/Log';
-import Navigation from '@libs/Navigation/Navigation';
-import TransitionTracker from '@libs/Navigation/TransitionTracker';
-import variables from '@styles/variables';
-import {setNameValuePair} from '@userActions/User';
-import CONST from '@src/CONST';
-import ONYXKEYS from '@src/ONYXKEYS';
-import type IconAsset from '@src/types/utils/IconAsset';
-import FeatureTrainingModalBody from './FeatureTrainingModalBody';
-import FeatureTrainingModalCarouselBody from './FeatureTrainingModalCarouselBody';
-
-const MODAL_PADDING = variables.spacing2;
-
-type BaseFeatureTrainingModalProps = {
-    /** The aspect ratio to preserve for the icon, video or animation */
-    illustrationAspectRatio?: number;
-
-    /** Style for the inner container of the animation */
-    illustrationInnerContainerStyle?: StyleProp<ViewStyle>;
-
-    /** Style for the outer container of the animation */
-    illustrationOuterContainerStyle?: StyleProp<ViewStyle>;
-
-    /** Style for the title */
-    titleStyles?: StyleProp<TextStyle>;
-
-    /** Whether to show `Don't show me this again` option */
-    shouldShowDismissModalOption?: boolean;
-
-    /** A callback to call when user confirms the tutorial */
-    onConfirm?: (willShowAgain: boolean) => void;
-
-    /** A callback to call when modal closes */
-    onClose?: () => void;
-
-    /** Text to show on secondary button */
-    helpText?: string;
-
-    /** Link to navigate to when user wants to learn more */
-    onHelp?: () => void;
-
-    /** Styles for the content container */
-    contentInnerContainerStyles?: StyleProp<ViewStyle>;
-
-    /** Styles for the content outer container */
-    contentOuterContainerStyles?: StyleProp<ViewStyle>;
-
-    /** Styles for the modal inner container */
-    modalInnerContainerStyle?: ViewStyle;
-
-    /** Children to show below title and description and above buttons (single-page mode only) */
-    children?: React.ReactNode;
-
-    /** Modal width */
-    width?: number;
-
-    /** Whether to disable the modal */
-    isModalDisabled?: boolean;
-
-    /** Whether the modal image is a SVG */
-    shouldRenderSVG?: boolean;
-
-    /** Whether the modal description is written in HTML */
-    shouldRenderHTMLDescription?: boolean;
-
-    /** Whether the modal will be closed on confirm */
-    shouldCloseOnConfirm?: boolean;
-
-    /** Whether the modal should avoid the keyboard */
-    avoidKeyboard?: boolean;
-
-    /** Whether the modal content is scrollable */
-    shouldUseScrollView?: boolean;
-
-    /** Whether the modal is displaying a confirmation loading spinner (useful when fetching data from API during confirmation) */
-    shouldShowConfirmationLoader?: boolean;
-
-    /** Whether the user can confirm the tutorial while offline */
-    canConfirmWhileOffline?: boolean;
-
-    /** Whether to navigate back when closing the modal */
-    shouldGoBack?: boolean;
-
-    /** Whether to call onHelp when modal is hidden completely */
-    shouldCallOnHelpWhenModalHidden?: boolean;
-
-    /** Sentry label for the help/skip button */
-    helpSentryLabel?: string;
-
-    /** Sentry label for the confirm/submit button */
-    confirmSentryLabel?: string;
-};
-
-type FeatureTrainingModalContentProps = {
-    /** Title for the modal */
-    title?: string | React.ReactNode;
-
-    /** Subtitle for the modal */
-    subtitle?: string;
-
-    /** Describe what is showing */
-    description?: string;
-
-    /** Secondary description rendered with additional space */
-    secondaryDescription?: string;
-
-    /** Text to show on primary button */
-    confirmText: string;
-};
-
-type FeatureTrainingModalVideoProps = {
-    /** Animation to show when video is unavailable. Useful when app is offline */
-    animation?: DotLottieAnimation;
-
-    /** Additional styles for the animation */
-    animationStyle?: StyleProp<ViewStyle>;
-
-    /** URL for the video */
-    videoURL?: string;
-};
-
-type FeatureTrainingModalSVGProps = {
-    /** Expensicon for the page */
-    image: IconAsset;
-
-    /** Determines how the image should be resized to fit its container */
-    contentFitImage?: ImageContentFit;
-
-    /** The width of the image */
-    imageWidth?: ImageSVGProps['width'];
-
-    /** The height of the image */
-    imageHeight?: ImageSVGProps['height'];
-};
-
-// This page requires either an icon or a video/animation, but not both.
-type FeatureTrainingModalIllustrationProps = MergeExclusive<FeatureTrainingModalVideoProps, FeatureTrainingModalSVGProps>;
-
-type FeatureTrainingModalPageProps = FeatureTrainingModalIllustrationProps & FeatureTrainingModalContentProps;
-
-type FeatureTrainingModalCarouselProps = {
-    /**
-     * When provided (and length > 1), the modal renders a horizontal paging carousel.
-     * The primary button advances to the next page until the last page, where it fires `onConfirm`.
-     */
-    pages: FeatureTrainingModalPageProps[];
-
-    /** Called when the user swipes to a different page. */
-    onPageChange?: (index: number) => void;
-};
-
-// Either single-page content fields OR carousel pages, but not both.
-type FeatureTrainingModalProps = BaseFeatureTrainingModalProps & MergeExclusive<FeatureTrainingModalPageProps, FeatureTrainingModalCarouselProps>;
-
-function FeatureTrainingModal({
-    illustrationInnerContainerStyle,
-    illustrationOuterContainerStyle,
-    illustrationAspectRatio: illustrationAspectRatioProp,
-    width = variables.featureTrainingModalWidth,
-    title = '',
-    subtitle = '',
-    description = '',
-    secondaryDescription = '',
-    titleStyles,
-    shouldShowDismissModalOption = false,
-    confirmText = '',
-    onConfirm = () => {},
-    onClose = () => {},
-    helpText = '',
-    onHelp = () => {},
-    children,
-    contentInnerContainerStyles,
-    contentOuterContainerStyles,
-    modalInnerContainerStyle,
-    isModalDisabled = true,
-    shouldRenderSVG = true,
-    shouldRenderHTMLDescription = false,
-    shouldCloseOnConfirm = true,
-    avoidKeyboard = false,
-    shouldUseScrollView: shouldUseScrollViewProp = false,
-    shouldShowConfirmationLoader = false,
-    canConfirmWhileOffline = true,
-    shouldGoBack = true,
-    shouldCallOnHelpWhenModalHidden = false,
-    helpSentryLabel,
-    confirmSentryLabel,
-    pages,
-    onPageChange,
-    ...props
-}: FeatureTrainingModalProps) {
-    const styles = useThemeStyles();
-    const StyleUtils = useStyleUtils();
-    const {onboardingIsMediumOrLargerScreenWidth} = useResponsiveLayout();
-    const {windowHeight, windowWidth} = useWindowDimensions();
-    const [isModalVisible, setIsModalVisible] = useState(false);
-    const [willShowAgain, setWillShowAgain] = useState(true);
-    const hasHelpButtonBeenPressed = useRef(false);
-    const pendingCloseRef = useRef(false);
-    const scrollViewRef = useRef<RNScrollView>(null);
-    const [containerHeight, setContainerHeight] = useState(0);
-    const [contentHeight, setContentHeight] = useState(0);
-    const insets = useSafeAreaInsets();
-    const {isKeyboardActive} = useKeyboardState();
-    const isInLandscapeMode = isInLandscapeModeUtil(windowWidth, windowHeight);
-
-    const isCarousel = !!pages && pages.length > 1;
-
-    const shouldUseScrollView = shouldUseScrollViewProp || isInLandscapeMode;
-
-    useEffect(() => {
-        // Transition tracker is used directly as we defer the opening of the modal until other animations are finished,
-        // for which there is no higher-level API.
-        const handle = TransitionTracker.runAfterTransitions({
-            callback: () => {
-                if (!isModalDisabled) {
-                    setIsModalVisible(false);
-                    return;
-                }
-                setIsModalVisible(true);
-            },
-        });
-        return () => handle.cancel();
-    }, [isModalDisabled]);
-
-    const toggleWillShowAgain = () => setWillShowAgain((prevWillShowAgain) => !prevWillShowAgain);
-
-    const pendingCloseModalAction = () => {
-        Log.hmmm(`[FeatureTrainingModal] Modal hidden - shouldGoBack: ${shouldGoBack}, hasOnClose: ${!!onClose}`);
-        if (shouldGoBack) {
-            Log.hmmm('[FeatureTrainingModal] Navigating back');
-            Navigation.goBack();
-        }
-        if (onClose) {
-            Log.hmmm('[FeatureTrainingModal] Calling onClose callback');
-            onClose();
-        } else {
-            Log.hmmm('[FeatureTrainingModal] No onClose callback provided');
-        }
-    };
-
-    const closeModal = (didPressHelpButton?: boolean) => {
-        if (didPressHelpButton) {
-            hasHelpButtonBeenPressed.current = true;
-        }
-        Log.hmmm(`[FeatureTrainingModal] closeModal called - willShowAgain: ${willShowAgain}, shouldGoBack: ${shouldGoBack}, hasOnClose: ${!!onClose}`);
-
-        if (!willShowAgain) {
-            Log.hmmm('[FeatureTrainingModal] Dismissing track training modal');
-            setNameValuePair(ONYXKEYS.NVP_HAS_SEEN_TRACK_TRAINING, true, false);
-        }
-
-        Log.hmmm('[FeatureTrainingModal] Setting modal invisible');
-        pendingCloseRef.current = true;
-        setIsModalVisible(false);
-    };
-
-    const closeAndConfirmModal = () => {
-        Log.hmmm(`[FeatureTrainingModal] Button pressed - shouldCloseOnConfirm: ${shouldCloseOnConfirm}, hasOnConfirm: ${!!onConfirm}, willShowAgain: ${willShowAgain}`);
-
-        if (shouldCloseOnConfirm) {
-            Log.hmmm('[FeatureTrainingModal] Calling closeModal');
-            closeModal();
-        }
-
-        if (onConfirm) {
-            Log.hmmm('[FeatureTrainingModal] Calling onConfirm callback');
-            onConfirm(willShowAgain);
-        } else {
-            Log.hmmm('[FeatureTrainingModal] No onConfirm callback provided');
-        }
-    };
-
-    // Scrolls modal to the bottom when keyboard appears so the action buttons are visible.
-    useEffect(() => {
-        if (contentHeight <= containerHeight || onboardingIsMediumOrLargerScreenWidth || !shouldUseScrollView || isCarousel) {
-            return;
-        }
-        scrollViewRef.current?.scrollToEnd({animated: false});
-    }, [contentHeight, containerHeight, onboardingIsMediumOrLargerScreenWidth, shouldUseScrollView, isCarousel]);
-
-    const Wrapper = shouldUseScrollView ? ScrollView : View;
-
-    const wrapperStyles = shouldUseScrollView ? StyleUtils.getScrollableFeatureTrainingModalStyles(insets, isKeyboardActive) : {};
-
-    return (
-        <Modal
-            avoidKeyboard={avoidKeyboard}
-            isVisible={isModalVisible}
-            type={onboardingIsMediumOrLargerScreenWidth ? CONST.MODAL.MODAL_TYPE.CENTERED_UNSWIPEABLE : CONST.MODAL.MODAL_TYPE.BOTTOM_DOCKED}
-            onClose={closeModal}
-            innerContainerStyle={{
-                boxShadow: 'none',
-                ...(shouldUseScrollView ? styles.pb0 : styles.pb5),
-                paddingTop: onboardingIsMediumOrLargerScreenWidth ? undefined : MODAL_PADDING,
-                ...(onboardingIsMediumOrLargerScreenWidth
-                    ? // Override styles defined by MODAL.MODAL_TYPE.CENTERED_UNSWIPEABLE
-                      // To make it take as little space as possible.
-                      {
-                          flex: undefined,
-                          width: 'auto',
-                      }
-                    : {}),
-                ...modalInnerContainerStyle,
-            }}
-            onModalHide={() => {
-                if (pendingCloseRef.current) {
-                    pendingCloseRef.current = false;
-                    pendingCloseModalAction();
-                }
-                if (!shouldCallOnHelpWhenModalHidden || !hasHelpButtonBeenPressed.current) {
-                    return;
-                }
-                onHelp();
-            }}
-            shouldDisableBottomSafeAreaPadding={shouldUseScrollView}
-            shouldWrapModalChildrenInScrollViewIfBottomDockedInLandscapeMode={!shouldUseScrollView}
-        >
-            {isCarousel ? (
-                <FeatureTrainingModalCarouselBody
-                    pages={pages}
-                    modalPadding={MODAL_PADDING}
-                    width={width}
-                    titleStyles={titleStyles}
-                    wrapperStyles={wrapperStyles.containerStyle}
-                    illustrationAspectRatio={illustrationAspectRatioProp}
-                    illustrationInnerContainerStyle={illustrationInnerContainerStyle}
-                    illustrationOuterContainerStyle={illustrationOuterContainerStyle}
-                    shouldRenderSVG={shouldRenderSVG}
-                    shouldRenderHTMLDescription={shouldRenderHTMLDescription}
-                    shouldShowDismissModalOption={shouldShowDismissModalOption}
-                    helpText={helpText}
-                    onHelp={onHelp}
-                    shouldCallOnHelpWhenModalHidden={shouldCallOnHelpWhenModalHidden}
-                    helpSentryLabel={helpSentryLabel}
-                    confirmSentryLabel={confirmSentryLabel}
-                    shouldShowConfirmationLoader={shouldShowConfirmationLoader}
-                    canConfirmWhileOffline={canConfirmWhileOffline}
-                    contentInnerContainerStyles={contentInnerContainerStyles}
-                    contentOuterContainerStyles={contentOuterContainerStyles}
-                    willShowAgain={willShowAgain}
-                    toggleWillShowAgain={toggleWillShowAgain}
-                    closeModal={closeModal}
-                    onConfirm={closeAndConfirmModal}
-                    onPageChange={onPageChange}
-                />
-            ) : (
-                <Wrapper
-                    scrollsToTop={false}
-                    style={[
-                        onboardingIsMediumOrLargerScreenWidth && StyleUtils.getWidthStyle(width),
-                        wrapperStyles.style,
-                        isInLandscapeMode ? {maxHeight: windowHeight * CONST.MODAL_MAX_HEIGHT_TO_WINDOW_HEIGHT_RATIO_LANDSCAPE_MODE} : styles.mh100,
-                    ]}
-                    contentContainerStyle={wrapperStyles.containerStyle}
-                    keyboardShouldPersistTaps={shouldUseScrollView ? 'handled' : undefined}
-                    ref={shouldUseScrollView ? scrollViewRef : undefined}
-                    onLayout={shouldUseScrollView ? (e: LayoutChangeEvent) => setContainerHeight(e.nativeEvent.layout.height) : undefined}
-                    onContentSizeChange={shouldUseScrollView ? (_w: number, h: number) => setContentHeight(h) : undefined}
-                    // Wrapper is either a View or ScrollView, which is also a View.
-                    // eslint-disable-next-line react/forbid-component-props
-                    fsClass={CONST.FULLSTORY.CLASS.UNMASK}
-                >
-                    <FeatureTrainingModalBody
-                        illustrationInnerContainerStyle={illustrationInnerContainerStyle}
-                        illustrationOuterContainerStyle={illustrationOuterContainerStyle}
-                        illustrationAspectRatio={illustrationAspectRatioProp}
-                        width={width}
-                        title={title}
-                        subtitle={subtitle}
-                        description={description}
-                        secondaryDescription={secondaryDescription}
-                        titleStyles={titleStyles}
-                        shouldShowDismissModalOption={shouldShowDismissModalOption}
-                        confirmText={confirmText}
-                        helpText={helpText}
-                        onHelp={onHelp}
-                        contentInnerContainerStyles={contentInnerContainerStyles}
-                        contentOuterContainerStyles={contentOuterContainerStyles}
-                        shouldRenderSVG={shouldRenderSVG}
-                        shouldRenderHTMLDescription={shouldRenderHTMLDescription}
-                        shouldShowConfirmationLoader={shouldShowConfirmationLoader}
-                        canConfirmWhileOffline={canConfirmWhileOffline}
-                        shouldCallOnHelpWhenModalHidden={shouldCallOnHelpWhenModalHidden}
-                        helpSentryLabel={helpSentryLabel}
-                        confirmSentryLabel={confirmSentryLabel}
-                        modalPadding={MODAL_PADDING}
-                        willShowAgain={willShowAgain}
-                        toggleWillShowAgain={toggleWillShowAgain}
-                        closeModal={closeModal}
-                        confirmModal={closeAndConfirmModal}
-                        {...props}
-                    >
-                        {children}
-                    </FeatureTrainingModalBody>
-                </Wrapper>
-            )}
-        </Modal>
-    );
-}
-
-export default FeatureTrainingModal;
-
-export type {
-    BaseFeatureTrainingModalProps,
-    FeatureTrainingModalProps,
-    FeatureTrainingModalCarouselProps,
-    FeatureTrainingModalContentProps,
-    FeatureTrainingModalPageProps,
-    FeatureTrainingModalIllustrationProps,
-};
