+++ src/components/Lottie/index.tsx
@@ -1,7 +1,6 @@
 import {NavigationContainerRefContext, NavigationContext} from '@react-navigation/native';
 import type {AnimationObject, LottieViewProps} from 'lottie-react-native';
 import LottieView from 'lottie-react-native';
-import type {ForwardedRef} from 'react';
 import React, {useContext, useEffect, useRef, useState} from 'react';
 import {View} from 'react-native';
 import type DotLottieAnimation from '@components/LottieAnimations/types';
@@ -16,12 +15,11 @@ import CONST from '@src/CONST';
 import {useSplashScreenState} from '@src/SplashScreenStateContext';
 
 type Props = {
-    ref?: ForwardedRef<LottieView | null>;
     source: DotLottieAnimation;
     shouldLoadAfterInteractions?: boolean;
 } & Omit<LottieViewProps, 'source'>;
 
-function Lottie({ref, source, webStyle, shouldLoadAfterInteractions, ...props}: Props) {
+function Lottie({source, webStyle, shouldLoadAfterInteractions, ...props}: Props) {
     const animationRef = useRef<LottieView | null>(null);
     const appState = useAppState();
     const {splashScreenState} = useSplashScreenState();
@@ -123,14 +121,8 @@ function Lottie({ref, source, webStyle, shouldLoadAfterInteractions, ...props}: 
             key={`${hasNavigatedAway}`}
             ref={(newRef) => {
                 animationRef.current = newRef;
-                if (typeof ref === 'function') {
-                    ref(newRef);
-                } else if (ref && 'current' in ref) {
-                    // eslint-disable-next-line no-param-reassign
-                    ref.current = newRef;
-                }
             }}
-            autoPlay={props.autoPlay && !isReduceMotionEnabled}
+            autoPlay={!isReduceMotionEnabled}
             style={[aspectRatioStyle, props.style]}
             webStyle={{...aspectRatioStyle, ...webStyle}}
             onAnimationFailure={() => setIsError(true)}
