+++ src/components/Section/index.tsx
@@ -166,7 +166,6 @@ function Section({
                                 style={styles.h100}
                                 webStyle={styles.h100}
                                 loop
-                                autoPlay
                                 shouldLoadAfterInteractions={shouldUseNarrowLayout}
                             />
                         ) : (
