+++ src/libs/Navigation/guards/index.ts
@@ -4,8 +4,7 @@ import type {OnyxEntry} from 'react-native-onyx';
 import getCurrentUrl from '@libs/Navigation/currentUrl';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {Session} from '@src/types/onyx';
-import AIFeaturesPromoGuard, {onSessionOrLoadingAppChanged as onAIFeaturesPromoSessionOrLoadingAppChanged} from './AIFeaturesPromoGuard';
-import MigratedUserWelcomeModalGuard, {onSessionOrLoadingAppChanged as onMigratedUserWelcomeModalSessionOrLoadingAppChanged} from './MigratedUserWelcomeModalGuard';
+import MigratedUserWelcomeModalGuard, {onSessionOrLoadingAppChanged} from './MigratedUserWelcomeModalGuard';
 import OnboardingGuard from './OnboardingGuard';
 import type {GuardContext, GuardResult, NavigationGuard} from './types';
 
@@ -20,8 +19,7 @@ Onyx.connectWithoutView({
     key: ONYXKEYS.SESSION,
     callback: (value) => {
         session = value;
-        onMigratedUserWelcomeModalSessionOrLoadingAppChanged(session, isLoadingApp);
-        onAIFeaturesPromoSessionOrLoadingAppChanged(session, isLoadingApp);
+        onSessionOrLoadingAppChanged(session, isLoadingApp);
     },
 });
 
@@ -29,8 +27,7 @@ Onyx.connectWithoutView({
     key: ONYXKEYS.IS_LOADING_APP,
     callback: (value) => {
         isLoadingApp = value ?? true;
-        onMigratedUserWelcomeModalSessionOrLoadingAppChanged(session, isLoadingApp);
-        onAIFeaturesPromoSessionOrLoadingAppChanged(session, isLoadingApp);
+        onSessionOrLoadingAppChanged(session, isLoadingApp);
     },
 });
 
@@ -106,6 +103,5 @@ function clearGuards(): void {
 
 registerGuard(OnboardingGuard);
 registerGuard(MigratedUserWelcomeModalGuard);
-registerGuard(AIFeaturesPromoGuard);
 
 export {registerGuard, createGuardContext, evaluateGuards, getRegisteredGuards, clearGuards};
